bank42/requirements.txt
Django >= 3.1.1, < 3.2.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
requests >= 2.25.1

